+++ comp/core/tagger/impl-mock/fake_tagger.go
@@ -3,21 +3,33 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package taggerimpl
+package mock
 
 import (
 	"context"
+	"net/http"
 	"strconv"
 
+	api "github.com/DataDog/datadog-agent/comp/api/api/def"
 	taggercommon "github.com/DataDog/datadog-agent/comp/core/tagger/common"
 	tagger "github.com/DataDog/datadog-agent/comp/core/tagger/def"
 	"github.com/DataDog/datadog-agent/comp/core/tagger/tagstore"
 	"github.com/DataDog/datadog-agent/comp/core/tagger/telemetry"
 	"github.com/DataDog/datadog-agent/comp/core/tagger/types"
-
 	taggertypes "github.com/DataDog/datadog-agent/pkg/tagger/types"
 	"github.com/DataDog/datadog-agent/pkg/tagset"
 )
+
+// Mock implements mock-specific methods for the tagger component.
+type Mock interface {
+	tagger.Component
+
+	// SetTags allows to set tags in the mock fake tagger
+	SetTags(entityID types.EntityID, source string, low, orch, high, std []string)
+
+	// SetGlobalTags allows to set tags in store for the global entity
+	SetGlobalTags(low, orch, high, std []string)
+}
 
 // FakeTagger is a fake implementation of the tagger interface
 type FakeTagger struct {
@@ -25,10 +37,20 @@ type FakeTagger struct {
 	store  *tagstore.TagStore
 }
 
-func newFakeTagger() *FakeTagger {
-	return &FakeTagger{
-		errors: make(map[string]error),
-		store:  tagstore.NewTagStore(nil),
+// Provides is a struct containing the mock and the endpoint
+type Provides struct {
+	Comp     Mock
+	Endpoint api.AgentEndpointProvider
+}
+
+// New instantiates a new fake tagger
+func New() Provides {
+	return Provides{
+		Comp: &FakeTagger{
+			errors: make(map[string]error),
+			store:  tagstore.NewTagStore(nil),
+		},
+		Endpoint: api.NewAgentEndpointProvider(mockHandleRequest, "/tagger-list", "GET"),
 	}
 }
 
@@ -169,4 +191,9 @@ func (f *FakeTagger) ChecksCardinality() types.TagCardinality {
 // DogstatsdCardinality noop
 func (f *FakeTagger) DogstatsdCardinality() types.TagCardinality {
 	return types.LowCardinality
+}
+
+// mockHandleRequest is a simple mocked http.Handler function to test the route is registered correctly on the api component
+func mockHandleRequest(w http.ResponseWriter, _ *http.Request) {
+	w.Write([]byte("OK"))
 }
